requirements.txt
streamlit
google-generativeai
python-dotenv
streamlit-option-menu